requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
